completed request handler
implimented request handler using a map that takes the path as a key and
a function as the value. That function takes a variable of type request
and outputs a variable of type response. It uses the newly created
serveFile function which takes in two strings, a filename / path, and a
status code and outputs a response stuct to pass the handler which
eventually will do error checking then will hand the data off to the
socket.